+++ Kindy/Kindy/Extensions/UILabel+.swift
@@ -1,0 +1,34 @@
+//
+//  UILabel+.swift
+//  Kindy
+//
+//  Created by WooriJoon on 2022/10/22.
+//
+
+import UIKit
+
+// UILabel의 행간격 조절 메소드 추가
+extension UILabel {
+
+    func setLineSpacing(lineSpacing: CGFloat = 0.0, lineHeightMultiple: CGFloat = 0.0) {
+
+        guard let labelText = self.text else { return }
+
+        let paragraphStyle = NSMutableParagraphStyle()
+        paragraphStyle.lineSpacing = lineSpacing
+        paragraphStyle.lineHeightMultiple = lineHeightMultiple
+
+        let attributedString: NSMutableAttributedString
+        if let labelattributedText = self.attributedText {
+            attributedString = NSMutableAttributedString(attributedString: labelattributedText)
+        } else {
+            attributedString = NSMutableAttributedString(string: labelText)
+        }
+
+        // (Swift 4.2 and above) Line spacing attribute
+        attributedString.addAttribute(NSAttributedString.Key.paragraphStyle, value: paragraphStyle, range: NSMakeRange(0, attributedString.length))
+
+        self.attributedText = attributedString
+    }
+}
+// 출처: https://stackoverflow.com/questions/39158604/how-to-increase-line-spacing-in-uilabel-in-swift
